Change DNS lookup to functions
We originally wanted and tried implementing the DNS lookup functions as
free functions but a compiler bug prevented that from working and so we
landed on using the DNS actor. That bug has been fixed, so we can now
replace the DNS actor with free functions!

We shouldn't need any worker thread pinning for this since we are just
doing a single query and thus there is only a single interaction with
the libuv loop.