Remove outdated out files
Remove files from the out/ directory that do not have a corresponding
source file in the src/ directory. This can happen if we have compiled a
project and then rename a source file. The old generate .c and .h files
would remain and the low level builder would find them and attempt to
compile them. This way we clean out such files without having to resort
to something akin of 'make clean'.