Fix altOutput when there are multiple modules
We would earlier exit after printing the alt output but this only works
when there's a single module. For multiple modules, like if we import a
dependency, only the first module would be printed.

Now all modules are printed and the module name is included in the
output too.

We take extra care not to run C compilation or similar to avoid any
failure in those steps stopping the overall compilation (like C
compilation of a dependent module would stop before we got to the
interesting module).