Optimize readyQ by adding tail pointer
Adding to the end of the readyQ is a really common operation and when
the readyQ grows this is taking longer and longer. By adding a tail
pointer, we optimize insertions and in particularly for long queues this
makes a tremendous difference.

The queue related code has also moved to its own source files so they
can be more easily tested. I have kept the old version guarded by the
MPMC variable (Multi-Produce Multi-Consumer), so this is version 2
currently. In some future I'd like to add an atomics based version 3 but
that's for later...