Optimize msg queue by adding tail pointer
Just like for the readyQ, we now keep track of a tail pointer so we can
add to the queue. When we add a lot to the same actors msg queue so the
queue grows in size this makes for quite a performance hit. On my AMD
5950x, running with a single worker thread, I get about 1.9 million
continuations per second running 1000 pairs of actors that ping-pong
each other async style so they have 1 outstanding message per pair. When
I increase to 500 outstanding tokens, the performance drops to about
350k continuations per second. With this tail pointer optimization, I
see about 1.8 million continuations per second, so a slight drop but
there's also some variation in the testing so don't read too much into
those numbers. Overall it's "much better with tail pointer" :)

The old implementation is left in, in case we want to perform more
systematic and accurate testing in the future. It's controlled by MSGQ.