Add mbedtls & tlsuv
tlsuv is a library that marries TLS with libuv, which is really
something we need since building TLS support on top of libuv directly is
non-trivial. For the TLS part it relies on an existing TLS library,
either OpenSSL or MbedTLS. I picked MbedTLS in this case since I could
get it to build with zig.

I did make an attempt at building OpenSSL based on a build.zig file I
found on the Internetz but it was incomplete and adding the missing
files that we need didn't really work as they in turn relied on some
bespoke build tool with a custom code generator in the OpenSSL repo. Way
too much yak shaving!

MbedTLS should be fine for now. Not quite sure how well it supports TLS
1.3 but there should be some basic support and they seem to be working
on it. Longer term I'd like to use the crypto stuff in zigs stdlib.
BoringSSL could be another alternative.