Add net.TLSConnection()
This adds a TLS client connection actor. It's interface is rather
similar to the TCPConnection actor.

Under the hood, we are using tlsuv and mbedtls for the TLS functionality
and integration with libuv. It is quite integrated so for example, at
this point it has name resolution integrated. That's a part I'd like to
lift out so we can do it in Acton (and one day avoid relying on
getaddrinfo entirely) but it's there for now out of simplicity. There's
more important issues to fix on this TLSConnection first, like disabling
certification verification (useful for testing).

Longer term I'd also like to see the ability to bundle CA certs in the
application itself so that we don't rely on anything external, think of
a WASM app or on a freestanding target.

There's a simple test application that tests our TLS client against
www.google.com. Since we don't have a TLS server and I don't want to
rely on external tools, like Python with some SSL module, this is sort
of the best we can do. As we don't want tests that rely on Internet
connectivity, this test is not run automatically.