Allow disabling TLS certificate verification
This is obviously not a good security practice but for the sake of
testing or similar it is very useful to be able to disable the
verification of TLS certificates.

Most of the changes went into the tlsuv library where I had to add new
options for this. I'm not all too pleased with that solution but it
works.

I have verified that this works by connecting to google.com but through
an IP address so we get a mismatch between the address and certificate
name. It normally fails but once we disable verification the connection
is established.