Add basic logging module
This adds our first attempt at a logging module for logging in an actor
centric world. I'm no logging expert, not like I've spent half my life
thinking of this and most of the experience I have is with imperative
language, perhaps Python above all. Python has lots of magic going on
with an implicit logger hierarchy behind the scenes that automatically
figure out modules etc. We don't have mutable state and I just don't
think it's a good fit. One of the benefits of Pythons implicit hierarchy
is the control per submodule or module tree. I tried to recreate the
same using Handlers in this module so that we can construct a hierarchy
of Handlers that give similar control to set output level for everything
below it in the hierarchy tree. The tree hierarchy is not implied
though, so the  developer has to explicitly define a hierarchy of
Handlers if that's what we want. More work, more control. Better? Not
sure. It's a bit of an experiment, let's see how we can use this.

We avoid loops between handlers by only allowing a single parent.

Not sure how fast this is. We do collect a bit of data for every
message, like timestamps (which is a syscall) etc etc but I think it
should be fast enough for the time being. Convenience over speed for
most use cases (up to a point of course) but good logging is so valuable
that I certainly think so for the level we're at here.

I wanted to make sure that log statements for log levels below the
current output level run fast, so that's specifically quite optimized.
The Logger object which runs in the same actor that wants to emit a log
message, keeps track of the output level of its handler and does not
emit them if they do not meet the output level. This information is
cached locally so we avoid doing a lookup (which is an actor method call
after all). This does introduce a slight delay of 100ms in propagation
of changes to the output level but I think that's manageable. Running
100k log.debug() is now still very fast when the output_level is INFO.
That said, we don't have generic lazy / delayed string interpolation for
log messages, so if we want dynamic content in the log messages
themselves that's... well that's not something we can do today. I think
the modern way of logging is to use structured data, so maybe we just
remove instance data completely from the messages. Like instead of
"Received 123 bytes of data", we just have "Received data" and a
structured data dict with {"bytes": 123} instead?

The log levels are brought in from the syslog RFC. I'd also like to
align on OpenTelemetry, since that's the newer kid on the block when it
comes to logging yet still a standard with good backing from the CNCF.
Would be nice to do both quite natively or at least with low impedance
mismatch.