Add argparse module
This adds a new module to the Acton standard library: argparse

As the name suggest, in particular for those with a Python background.
argparse is about parsing command line arguments.

This module implements the basic concepts of positional arguments,
options with values and boolean flags. There are also commands. Compared
to the python argparse module, which is ambiguous for a lot of
configuration, we've tried to be less ambiguous. For example, it is not
possible to have multiple positional arguments expecting multiple
values, i.e. nargs=+ or nargs=*. For optional values that takes multiple
values, we require that the option is provided multiple times, like:

  --foo A --foo B

Whereas in Python, it is possible to do:

  --foo A B

But it again becomes ambiguous in combination with positional arguments,
which is why we avoid this.

We keep track of types and provide getter functions to read the values
with a guaranteed return type which is a little bit more idiomatic
Acton since we want to rely on our stronger types when we can.