Add chunked transfer-encoding & headers
This improves the HTTP library. First is a refactoring of the parsing
functions for parsing HTTP requests and HTTP responses. They now have a
common base in parse_message() which parses a HTTP message. Requests and
responses look the same except for the start line, so parse_message()
does the parsing and hands over the start_line so that parse_request()
and parse_response can parse just the start line.

There's now a test module which tests the parsing of requests and thus
of the main parse_message() function.

Now with the new parsing organization and tests in place, I've added
support for chunked transfer-encoding.

All the get, post, etc methods on the HTTP client now take a headers
argument so it is possible to send in headers.