AbE: fix bool field in tuple
Move run-example.py so it can be reused for multiple books, not just AbE.

Fix tuple example to directly check bool type as a bool. We fixed this
bug but I couldn't get the example to work, which turned out to be
because I was using the system installed actonc rather than the one
locally built in the repo, and so I ended up with an older one that
didn't have the fix.

I changed the path to actonc in run-example so we get the locally built
one which is probably what we want!?

Now also have a test on this type of bool check on a field in a returned
tuple.