Upgrade to Stack LTS 21.13 & GHC 9.4
It's been awhile since we upgraded and it would be nice to move to a
fresher version of ghc, so here we go. Some tests shows that actonc can
now run ~25% faster. Pretty cool!

GHC 9 requires a newer version of stack, so we run a stack upgrade on
the older distributions we test on CI.

Somewhat surprisingly, C++ now seems to be used for something. We must
unset CXX to avoid colliding with zig c++ and g++ must now be installed
on CI machines. It would be really nice if we could use zig cc and zig
c++ but alas, that doesn't work out of the box and appears tricky to fix.