Statically build actonc on Linux
We now do static compilation of actonc on Linux. Static compilation is
only supported on Linux so we condition this on the os being Linux.
Previously we would only build statically in CI for release builds
during the build-debs job, which was run on Debian:11. Building
statically would not work on newer distributions due to this glibc issue
around __xmknod that popped up with glibc 2.33 (I think). After
upgrading stack LTS & GHC, we can now statically compile on any version
of Linux so that's what we do. Not that static binaries make for much of
an improvement for local builds, but it doesn't hurt either.

So -static as a ld-options is apparently the correct incantation.  The
previously used ghc-option -optl-static is bad somehow, because we tell
ghc but not the linker, I guess. Internet can be misleading.

Anyhow, that alone doesn't work due to this issue with the __xmknod
symbol not being found when linking with glibc. It's the haskell 'unix'
package that wants to run mknod but I don't know why it ends up
resolving to __xmknod which is IIUC an internal symbol. This symbol is
only in the shared dynamic glibc, so we don't find it when trying to
statically compile. I've toyed around with this and half out of luck I
found that compiling with -fPIC seems to fix it. Yay. I also tried doing
static compilation of all dependencies or just the unix package but that
breaks things because I believe the flag propagates to dependencies of
the unix package and those in turn spew out some error. An alternative
that does work is to manually pull down the unix package, do a cabal
v2-configure --disable-shared, reference this new package from
stack.yaml and set some extra-deps correctly. However, seeing how that
is a much more elaborate process, I've opted for this one, although I
don't fully understand why or how it works.