Add __str__() method to Exceptions
The base exception now has a __str__() method that formats the
exception. This is very useful for more advanced exception types where
we want to be able to include more dynamic data but not carry the cost
of doing up front string formatting. Now we can always provide extra
data but only string formt it when necessary for printing. More complex
exception types need to implement their own __str__().

_name() returns the name of the exception type.

Now we can standardize on calling this __str__() function when printing
exceptions too!