Add functions to enable/disable GC + get mem usage
Also changed the gc() function to not only run a collection round but
also unmap free memory. From a few quick test runs it doesn't seem to be
working very well though, I wonder if it's because of memory
fragmentation?

get_mem_usage() is approximate. Would be nice if it was exact but I can't
seem to find underlying counters that are exact, so here we are.