Expect proc() in argparse cmds
() means pure functions which is not what we want here, we want to allow
wider effects, so proc() is right (we don't want actions since they'd be
run async - weird!).

When using this I have to do the argument parsing in an actor. I'm not
quite sure why but I can't seem to be able to do it using a plain function.