Add testing module
This adds a testing module which helps users to test their Acton code!

There's currently only one type of tests - "unit tests" - that are
supported. We define them as test functions that take no arguments and
return nothing (well None) when things go well. Tests should use the
testing.assert* functions to test whether various values are as expected.
Exceptions are used to signal when such values are not as expected.

We separate between test failures and errors. Failures are when a test
runs to completion but the outcome is not as expected. Test errors are
when there is some other error while running the test. Technically, this
is done by looking at the type of exception thrown. All test failures
should inherit from testing.AssertError. All other exceptions indicate a
test error.

Test discovery is currently not implemented. The idea is that the
compiler will find all functions in a module with a name prefix of
_test_. The type signature should match proc() -> None, and if so it
will be placed in a dict called __unit_tests(). Further, if any tests
are found, we should generate a new root actor entry point called
__test_main(env), which the compiler should

It should be possible to manually define a __test_main() actor in case
one wants to override the tests.

We should add support for other types of tests, probably providing env
to do IO but we're also then faced with how to structure the control
flow of the test itself since any IO implies a callback based structure.