Fix error handling in TCPConnection
We are waiting for an error both on IPv4 and IPv6 but since we
incorrectly reported IPv6 errors as IPv4, we never got both and thus
never propagated errors to the calling code via the on_error callback.

This only deals properly with dual stack connections. For IPv4-only or
IPv6-only cases we need more code paths...