Prefer plain env.exit()
Exiting an Acton application is asynchronous and this can be a little
surprising.

As env is an actor and env.exit() is a method on that actor, it will be
called asynchronously per default since there is usually no assignment
happening of the return value. There really is no return value either.

Most people, including myself, has the intuition that a program should
exit immediately when exit() is called and any line after it in the
program source should not be executed. This does not happen with a plain
env.exit() and so all examples and documentation was updated to use

  await async env.exit(0)

However, that's pretty ugly and so the next step was to change
env.exit() to be a builtin function instead, like print() or similar.

Now, another part of being a method on the env actor is that only actors
that hold a reference to env can actually call env.exit(). Anyone can
call a builtin function however, so to uphold our capabilities based
security, the builtin exit() would have to take a capability argument,
so the caller could prove privileges to exit the application, thus it
becomes something like:

  exit(env, 0)

That's not very pretty either and when discussing this code (I actually
did implement it), we finally decided to discard this idea altogether.

On program shutdown we want to be able to hook into the stop event and
run arbitrary code. In other imperative languages this can be done with
at-exit handlers and similar, which can run functions on exit. In Acton,
we run actor continuations and we rely on the RTS to run those
continuations. We can't have anything akin to at-exit handlers that just
run some functions. We would want to use actors to implement on-exit
logic too, so we need a different type of mechanism altogether. Thus,
the whole idea of doing a synchronous exit does not really rhyme well
at all with an actor world. Rather, we want to focus on the plain
env.exit(0) and rather instill in our users the idea that it is an
asynchronous process and to structure their code accordingly.