Change JSON encode/decode to free functions
The JSON encode and decode functions were implemented as methods on an
Actor due to an actonc bug where NotImplemented simply wasn't recognized
for free functions. This design was never the intended one but a side
effect of this actonc limitation. The json.encode() function
instantiated the Json actor and called Json.encode() so the
functionality of calling the encode function was possible just that it
involved creating an actor and so required quite a bit of extra overhead
than necessary.

Another consequence was that the encode() function got the effect proc,
since it creates an actor, and of course any calling could would be CPS
split. The unit testing functionality offered by the testing module
requires mut() effect so we can't use it to test JSON encoding or test
any function that in turn deals with JSON.

So, now I've finally swapped these around so the real encode/decode
happens in the free functions encode() & decode() in the module. The
Json ator still remains but now act as a thin outer shell for doing
async style processing of JSON.