Update changelog: key in KeyError.__str__()
I've done a simple performance benchmark of this. Doing a for loop with
a million iterations and trying to do a key lookup of something that
doesn't exist, thus raising a KeyError. It runs in about 1.15 seconds on
my machine from Acton main branch. Switching to this branch with the new
KeyError results in the same time, i.e. ~1.15 seconds. Also adding a
`str(exc)` in the loop to format the exception brings up the total time
to ~2.8 seconds, so quite a dramatic increase. Looks like everything is
working as it should!