Fix net.TCPConnection single stack error handling
One of the main features of the new TCPConnection is that it does rather
nice dual stack support. However error handling even presumed dual stack
and did not work correctly for single stack use, such as when an IP
address is given as input. That's now fixed.

The full state machine needs to be modeled and reviewed. This is still a
bit haphazardly put together and I'm sure there are more bugs lurking.

We've been seeing quite a lot of problems in CI with test timeouts. Not
sure why there would be timeouts or any errors in CI, but if they're
actually connection refused but without the error propagating, this
could explain it. The rate is suspiciously high though. Let's see.