Revert "Use per-thread malloc"
This reverts commit 1f7fcf3db5bf17ce988803edd29cfff585c12927.

The per thread malloc seems to cause hangs, so we need to disable it for
now and bring it back in in an order fashion.

I've been running the rts_db/test_tcp_client application repeatedly and
it keeps hanging in about 25% of cases with per thread malloc whereas
it always works without it.