Add type()
Johan Nordlander hates type() - it should not be used for
metaprogramming... but I find it damn helpful as a debug tool. I print
types and work with that :) Maybe a "print_type" could fix this, which
directly prints to stdout instead of returning a str, that way it's
impossible to use for metaprogramming but anyhow, let's go for this.
We'll try to remove it in the future when we have other better means of
debugging types, like a language server or similar.