Release v0.18.2
Keep posting deb updates to the bullseye APT distribution. It failed
because we haven't initiated the bookworm one. I mistook the error to
mean a misalignment between what we tried to add to and the source (i.e.
what is in changelog), but that was wrong.

Our packages are compatible across bullseye and bookworm. In a perfect
world we'd probably add both but it feels simpler and more pragmatic to
just use one and between calling it bookworm or bullseye, I think it's a
better intuition to keep the old name since it is indeed compatible with
bullseye too. Easier to explain how it is forwards compatible.