Improve DNS error handling
If we get a DNS resolution error, we now check if we are still in the
resolving state. If not, like if we have already connected or are in
some other state where this is not relevant any longer, we don't do
anything. If we are still in the resolving state, we retry the DNS
lookup after 0.1 seconds.

Previously we would just retry immediately, which would lead to loads of
busy work. For example, if there is no AAAA record at all, we would just
continue retrying the AAAA lookup for all future, so even when the IPv4
connection has established and TCP is working properly, the TCP actor
would just retry the AAAA lookup. This could consume 50% - 75% of a CPU
core!