Add more TODOs for limiting auto reconnects
Add some basic heuristics, like when we create a http.Client the first
time we obviously connect directly then we expect to run at least one
query. Maybe we should not reconnect after that (in case of HTTP / 1.0
or not having persistent in later HTTP versions) until there is a second
query. Now if we see a second query we can assume "there are multiple
requests" and thus reconnect directly after the 2nd and subsequent
requests.