Support no-threads RTS mode
This adds a mode for running the RTS without threads. It can be useful
on platforms that do not have POSIX threads available, such as
WASM (although we are far from being able to support WASM for other
reasons).

Since one of the goals is to remove the need for pthreads, this must be
chosen at compile time rather than run time. Use the --no-threads flag:

  actonc --no-threads foo.act

The compilation of RTS, which is part of libActon, is affected by this
flag and so it must be recompiled with the flag set, thus this
invalidates prebuilt libActon and instead compiles it. It's pretty quick
though but still noticeably when not in a project (that caches).

The DB support currently requires threads. To get past compilation,
pthread is made optional and instead results in an error if attempted to
be used at run time.