Add Windows support
We don't support threads on Windows right now so compiling for Windows
implies disabling threads.

There are a number of regressions, like we disable certain functionality
to make programs compile for Windows, but the end-result seems to be a
functioning binary!

logging.ERROR had to be renamed as on a C-level, it becoms 'ERROR' which
seems to be some macro on Windows!?

We test cross-compile to aarch64/x86_64-windows-gnu!