Test external project Telemetrify
Telemetrify is the largest project implemented using Acton, by quite a
fair margin. We want to be aware if / when we introduce backwards
incompatible changes in Acton that breaks Telemetrify.

We test the acton-next branch, which should conceptually be the same as
the main branch but with the possibility of adding extra fixes to make
it compatible with acton head.