Add acton frontend + test cmd
This adds the new acton frontend which is meant to be our new primary
frontend rather than invoking 'actonc'. It's mostly a drop-in
replacement for now, like `actonc build` with all the option it takes
can be replaced with `acton build`. That all works except for `acton
new` where some bug in our argparse module prevents things from working
correctly.

Besides adding this new frontend, we also support a new command: `acton
test` which besides building the Acton project will run all tests. This
is quite complex as we need to invoke multiple programs (one for each
module) and receive and summarize the output. It's working although I
think there are some improvements to make!