Hardcode debian dependency to libc6 2.27
${shlibs:Depends} is a macro that gets expanded to the dynmically loaded
run time dependencies. The new acton frontend is compiled with --target
...-gnu.2.27 to depend on an older glibc, thus making it forward
compatible. The debian build system inspects the built binaries of a
package to determine the run time linked libraries but it cannot
differentiate between versions of glibc, so it assumes the glibc version
of the build machine. We know better, which is why we hard code it to
the older 2.27 version.

Statically linking acton could also be an option but I think there are
some consequences, like exec() and friends which we do use in the Acton
process stdlib module (and which the acton frontend makes extensive use
of!) might not work well.