Use zig 0.12.-dev.2236
There's been some major updates to the build system in zig and this
brings us up to date.

What's pretty cool is that we've done a proper job of reimplementing the
CMake build of BDWGC using build.zig (earlier it was just a quick hack
supporting a minimal subset of features) and it's now upstreamed, so the
git ref we're now using is from the upstream repo and not our own fork,
although we continue to mirror the repo for the time being.

Our installation size will grow now since we're also enabling C++ and
Windows, which we've previously stripped out from our zig install. This
is mostly a preparatory move.