actonc: line buffered stdout
Haskell detects if stdout is a terminal, in which case it does line
buffering, or something else in which case it is block buffered. This
means when we call it from the `acton` frontend we end up with block
buffered output which locks really bad. We want direct output which
gives that interactive feeling!