Let main thread perform special actor work
This adds so the main thread runs the worker, i.e. it picks up and
executes actor continuations. We've previously used readyQ 0 as the
shared queue, but since the main thread is the closest thing we have to
a thread 0, it now owns readyQ 0. The shared readyQ is instead at the
top-most position.

For now, only the Env actor runs on the main worker thread. This is
necessary since things like installing signal handlers, which we want to
do in the Env actor, must be performed in the main thread. There are
probably other good uses of placing particular actors on this thread.