Rename uncompress -> decompress
IMHO decompress is the verb / act of taking compressed data and
decompressing it. The resulting data is uncompressed, but uncompress can
hardly be a verb, so I think the method should be called decompress(),
thus the rename.

First sentence in Snappy README is "A fast compressor / decompressor",
not "uncompressor". I do not that thsnappy function is called
snappy_uncompress() though... but ugh.