argparse: better (proper?) nested command parsing
argparse now supports nested commands with a better algorithm for
finding args and what command they are associated with. We do a sort of
depth first search, whenever we encounter a command, we pass the rest of
the argv to that command parser and let it pick the arguments it wants.
Only after it is done do we take a second iteration over the rest of the
args to look for positional arguments.

This makes it possible to support optional positional arguments in a
parent command parser where a child parser takes a positional argument.
The acton cli frontend is an example of this. It is possible to run
`acton foo.act` to compile `foo.act`, where `foo.act` is an optional
positional argument. It is optional, otherwise `acton build` would give
an error. For `acton new my-project`, `new` is a sub-command that takes
a positional argument (the name of the project to create). Previously,
the parser would consume `my-project` as the positional argument to the
top level parser rather than passing it to the command parser for `new`.
This is now reversed. This new approach seem like a much more sound
algorithm.