Support run-time reconfigurable malloc
All calls to malloc & friends in builtins etc are now replaced with a
calls to acton_malloc & friends, which makes it possible to reconfigure
which malloc to use at runtime. Since we do link time redirection of
malloc, we use dlsym to dig out the real malloc. This makes it possible
to place things outside of the GC heap.