Support custom malloc
We have thus far relied on link time redirection of malloc and friends.
It's a little cleaner and more flexible to explicitly call GC_malloc or
whichever malloc we want. That way we can still use the classic malloc
if we want to and mix and match between different allocators.

acton_replace_allocator will now also change the allocator in all
relevant C libraries we make use of in stdlib and similar. libuv and
tlsuv are not affected, they always use GC_malloc, set from
acton_init_alloc() since nothing in module initialization that is
constant should come from these libraries and we want all libuv stuff to
be GC_malloced, so that all things in the IO loops can be properly
traced from the root set.

We're still doing link time redirection but this patch should put the
necessary changes in place to actually stop redirecting malloc. We need
to verify with testing before we flip the switch though.

Further, this patch actually fixes bugs, like for our int where the
initial allocation is done in int.c using malloc_int. During module
initialization, this points to real_malloc but if such an integer grows
beyond a single limb, it is reallocated from within the BSDNT zz.c file
which previously was hard-coded to use malloc, which is redirected to
GC_malloc(). Since the initial malloc was done using real_malloc(), the
object lacks the GC header and the subsequent GC_realloc() crashes. Now
with this patch, we actively set the BSDNT realloc function to point to
real_realloc during module initialization so it works.