Fix RE malloc
This was just plain wrong. We didn't notice because the stdlib tests
were not running after the restructuring there. I was running the tests
locally for a while but since they were not invoked by `make test` we
didn't notice the breakage of the re module.

Also uses a better xmlns URI. We get a warning printout for other values
which might trip up the test system (it shouldn't, but hey...).